requirements_backup.txt
requests
dnspython
python-whois
python-nmap
